10.13.4

+++ bind9/bind9/bin/tests/system/rsabigexponent/prereq.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 #
-# Copyright (C) 2012, 2014  Internet Systems Consortium, Inc. ("ISC")
+# Copyright (C) 2012, 2014, 2016  Internet Systems Consortium, Inc. ("ISC")
 #
 # Permission to use, copy, modify, and/or distribute this software for any
 # purpose with or without fee is hereby granted, provided that the above
@@ -17,12 +17,13 @@
 SYSTEMTESTTOP=..
 . $SYSTEMTESTTOP/conf.sh
 
-test -r $RANDFILE || $GENRANDOM 400 $RANDFILE 
+test -r $RANDFILE || $GENRANDOM 400 $RANDFILE
 
-if ./bigkey > /dev/null 2>&1
+if $BIGKEY > /dev/null 2>&1
 then
     rm -f Kexample.*
 else
-    echo "I:This test requires that --with-openssl was used." >&2
+    echo "I:This test requires cryptography" >&2
+    echo "I:configure with --with-openssl, or --with-pkcs11 and --enable-native-pkcs11" >&2
     exit 255
 fi
